Re-organized the project
Rather than having a subdirectory with the working code, and some
resources in the root of the git, I moved all the project files up to
the root directory and removed the resources that weren't needed. This
makes for cleaner looking project as well as a project that is more
easily transfered to the final repo when submission for grading is
necessary.