Lots of additions
Moved some structs from cub3D to engine since they're required there.
Had to rework how images are initialized and used.
should still re-work them a little. Right now I only make 2 but we
should probably put this in an array so I can use as many as I'd like.
Also created a paint bucket function that will fill a box with a color.
Also made a color edit function that edits the color of a specific pixel
in an image buffer.