Update to ft_draw_background
draw background takes an image and the engine (which contains the
image).
But I wanted to keep these as 2 separate arguments so other images can
be used if desired.
Changed the call to mlx_put_image_to_window so that it doesnt find the
image through the engine struct but rather uses the one provided in the
arguments.