Lots of changes!
Removed background.c and map.c because I dont need to prerender the map
or background, i need to prerender the images that will be used.
Added to image_id enum.
Added rect struct that contains top left coord, bottom right coord and a
color.
re worked set px color function and paint bucket function to take new
arguments.
Added call to new create rect image.
This new function will create an image, put it on the end of the image
linked list, and initialize it, and fill in the image with a color.