Added map_data->map_block_size
in the map_data struct, I added a varaible called map_block size so
calculating the dimentions of the map would be easier. Made the
calculations for this at the same time in the map parsing when the
map_data->size is determined.
Also added a drawing function for the map, but not finished yet.