Changed the player movement do direction based.
Now the player coordinates are not just added to. Rather we calculate
the player's rotation and control the player's delta.
Delta is the amount the player will move in the next step.
angle can be changed with A and D keys. W and S are used for movement
along this angle.