Re-worked the ft_draw_rays function
Now, rather than drawing each pixel for the rays directly to the window
(which takes a lot of overhead), I create an image and edit this image
before pushing it to the screen for each screen refresh.
Also added a ft_draw_map_rays which does the same but specifically for
the mini map. To reduce overhead, we could remove this function from the
render.
Now ft_draw_rays will call 3 functions, one for the ceiling (ie vertical
offset) one for the rays themselves, and one for the floor (ie the rest).
Also adds a ft_clear_image function which sets all pixel values in an
image to 0 (ie black). This may not be necessary either.
Also changes ft_init_ray so that the offsets are based on the FOV. This
way we always cast WIN_X number of rays and fit them inside of a FOV
size viewport.
We also remove the BG_IMAGE_F and BG_IMAGE_C from the created images in
prerender because these are no longer needed.
Sorry, this commit should have been smaller.