Running is smoother if map is large.
There is an issue where the coordinates of the player are directly tied
to how large the map is. So to get this to work, I would need to
re-write the ray casting algo.

Rather than doing this, I'm going to make the map large and hide it so I
can do the mando.