Basic texture mapping.
Lots of small changes as I tried to get this to work.
Final product is the first column of colors in the indicated texture
maps to the correct block, but no texture scaling and only one y texture
column.
Moved a few functions around and added a colors.c file for color related
functions.
Added back the img_buf init for xpm images since I need that for mapping
colors.
Height and width of the texture is currently hard coded! need to make
dynamic.
Large edit to ft_draw_wall as we map the colors from the texture to the
wall.