Small fixes.
Removed the ft_clear_image call which reset the ray image.
Turns out this was making things look bad and by removing it, some of
the previous frames can remain in the image. Since the sky is a constant
color, this is actually not an issue.
Also made the texture_x offset in draw_walls an absolute value of the
ray x or y coordinate. This way we dont accidentally index a negative
value which can cause a segmentation fault.